Fix typeform tests
